#  Add specialization and optimize