--- a/askama_escape/build.rs
+++ b/askama_escape/build.rs
@@ -1,0 +1,11 @@
+extern crate version_check;
+
+fn main() {
+    match version_check::is_nightly() {
+        Some(true) => {
+            println!("cargo:rustc-cfg=askama_nightly");
+        }
+        Some(false) => (),
+        None => (),
+    };
+}