@botika 2018-12-03
How about changing the filter `safe` from function to reading in compilation? In order to use the `MarkupDisplay` only when necessary. With it we will avoid the `match` every `fmt`.